web-ui/requirements.txt
flask>=3.0.0
flask-socketio>=5.3.6
paho-mqtt>=2.0.0
python-engineio>=4.9.0
python-socketio>=5.11.0
requests>=2.31.0
